BIG CHNAGES
Added a lot of plugins and functions for QoL. Necessary repos are in the links found in the plugin section.

I organized things somewhat, but there is always room for improvement. Still a lot of work to do but now I can focus more on coding and adapting to nvim instead of configuring it. Fugitive is there for git, and Im using lsp and treesitter for languages.

COC is used for a lot of the autocompletion, but I dont have a lot of languages installed. I need to work on a linter at some point.

I am not using telescope because for some reason i am allergic to lua. If a plugin relies on lua and not init.vim then i skipped it oops.

3/6/24 @ 8:29pm